Trait collator_selection::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type StakingCurrency: ReservableCurrency<Self::AccountId>;
    type RewardsCurrency: ReservableCurrency<Self::AccountId>;
    type UpdateOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type PotId: Get<PalletId>;
    type MaxCandidates: Get<u32>;
    type MinCandidates: Get<u32>;
    type MaxInvulnerables: Get<u32>;
    type KickThreshold: Get<BlockNumberFor<Self>>;
    type ValidatorId: Member + Parameter;
    type ValidatorIdOf: Convert<Self::AccountId, Option<Self::ValidatorId>>;
    type ValidatorRegistration: ValidatorRegistration<Self::ValidatorId>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Overarching event type.

type StakingCurrency: ReservableCurrency<Self::AccountId>

The currency used for staking.

type RewardsCurrency: ReservableCurrency<Self::AccountId>

The currency used for rewards.

type UpdateOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that can dictate updating parameters of this pallet.

type PotId: Get<PalletId>

Account Identifier from which the internal Pot is generated.

type MaxCandidates: Get<u32>

Maximum number of candidates that we should have. This is enforced in code.

This does not take into account the invulnerables.

type MinCandidates: Get<u32>

Minimum number of candidates that we should have. This is used for disaster recovery.

This does not take into account the invulnerables.

type MaxInvulnerables: Get<u32>

Maximum number of invulnerables. This is enforced in code.

type KickThreshold: Get<BlockNumberFor<Self>>

type ValidatorId: Member + Parameter

A stable ID for a validator.

type ValidatorIdOf: Convert<Self::AccountId, Option<Self::ValidatorId>>

A conversion from account ID to validator ID.

Its cost must be at most one storage read.

type ValidatorRegistration: ValidatorRegistration<Self::ValidatorId>

Validate a user is registered

type WeightInfo: WeightInfo

The weight information of this pallet.